requirements.txt
google-genai>=0.8.0
PyQt6>=6.4.0
psutil>=5.9.0
numpy>=1.24.0
sounddevice>=0.4.6
pyautogui>=0.9.54
playwright>=1.40.0
pillow>=10.0.0
pyperclip>=1.8.2
requests>=2.31.0
